requirements.txt
# Authlib==1.2.1
# boto3==1.26.97
celery==5.4.0
coverage==6.5.0
daphne==4.1.2
dj-database-url==2.3.0
Django==5.1
django-celery-results==2.5.1
django-cors-headers==4.2.0
django-filter==23.3
django-jazzmin==3.0.1
django-json-widget==2.0.1
django-jsonform==2.15.0
django-rest-passwordreset==1.3.0
django-simple-history==3.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
drf-yasg==1.21.8
gunicorn==23.0.0
jwskate==0.11.1
pillow==10.0.0
postmarker==1.0
psycopg2-binary==2.9.10
pyjwt==2.10.1
Pygments==2.19.1
qrcode==8.0
redis==5.2.1
sentry-sdk==2.20.0
sshkey-tools==0.11.3
# typing-extensions==4.5.0
web3==7.7.0
whitenoise==6.8.2
channels==4.2.0
django-multiselectfield==0.1.13
django-json-widget==2.0.1